// frontend/types/experimentList.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * My API
 * Documentação automática da API
 * OpenAPI spec version: 1.0.0
 */
import type { DesignTypeEnum } from './designTypeEnum';
import type { StatusEnum } from './statusEnum';

/**
 * Serializer simplificado para listagem de experimentos.
 */
export interface ExperimentList {
  readonly id: number;
  /** @pattern ^[-a-zA-Z0-9_]+$ */
  readonly slug: string;
  /** @maxLength 200 */
  title: string;
  design_type?: DesignTypeEnum;
  status?: StatusEnum;
  readonly owner_email: string;
  readonly owner_name: string;
  readonly created_at: string;
  readonly updated_at: string;
}